Allow for pathed wrapped commands
My environment provides a wrapper that points to a Ruby interpreter at a specific path. The path is quoted to prevent the usual issues around spaces in the path name. As a result, shebang-set-grammar fails to detect the language.

This change handles both situations. It handles quoted path by adding both double and single quotes to the first split. It then ignores any 0 length strings. It then detects the presence of "/" in the command string and returns the actual wrapped command without the path.

As a result, I can now open a file and Atom detects the grammar correctly, making me a much happier coder.